backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
pandas>=2.1.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
google-generativeai>=0.7.0
google-ai-generativelanguage>=0.6.0
